requirements.txt
mcp
pandas
requests
PyJWT
python-dotenv
